requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
google-genai
python-jose==3.5.0
requests==2.32.5
pydantic==2.12.5
python-dotenv==1.0.0
sqlalchemy
passlib[bcrypt]
python-multipart
python-dotenv
chromadb
redis